Improve bin/gen-years.sh
The links in the html file now are a bit more consistent in form and
the titles make a bit more sense.

By that I mean that it used to be that it would say 'About the 1st IOCCC' but only '1st IOCCC' was a link
and 'About' was bold but now it's all bold and now it's all a link (to suggest that the link is actually about
the first IOCCC). The download links were also improved in this way and
they now say 'winning entries' rather than say (for 1984) '1984 entry
source' (that was for all entries). Also, since it would otherwise say
something like:

	...
	Download all 1985 IOCCC winning entries.
	Download all 1984 IOCCC winning entries.

I reworded it to say:

	...
	Download all IOCCC winning entries of 1985.
	Download all IOCCC winning entries of 1984.

As other link titles were in bold in the file now each entry link is in
bold as well and so is the title - to help it stand out a bit better. I
am not quite sure about the bold title yet but the link should be bold
anyway to match others. I think the bold title does seem like a good
idea though.

Rebuilt years.html.

Note that most of the links have been checked in this file but I will
look at each entry link again as I verify links and formatting in the
entries' html files.