Minor fixes in bin/README.md
A few typos were fixed.

One tool had the description reworded a bit to make it (hopefully) a bit
clearer.

A link was fixed but I am not sure if, since it's a script, it should
link to the GitHub repo or not, like with other markdown files. I
changed it to use the %%var%% that the bin/README.md file uses. This is
one reason links have not been checked. The other reason is I have not
gone through the entire file yet.

I notice some tools do not have example usage and those could be added
too.

There was an apparent typo in bin/find-missing-links.sh and this same
typo was fixed in bin/README.md. I am not sure exactly what was meant,
however, so it might need further modifications.

I have (the caveats above notwithstanding) gone up through
pandoc-wrapper.sh or just before it.

Regenerated bin/index.html.