Finish format/typo checks in next/
I was asked to actually make corrections in the rules and guidelines and
to make sure that the look and feel seems good. The typos and other
features will be added back later. Some things worth noting below.

First references to twitter were changed to mastodon. I also in some
cases (this should be in the FAQ too and I'm not sure if it is) when
referring to mastodon noted that unless you are mentioned (also if
someone reposts your toot but this is not mentioned as it's not
relevant) you will not get a notification so one should still refresh
the page as otherwise they might not get an update.

The current iocccsize in the mkiocccentry repo checks for 2503 not 2053
like it used to. Only some of the references can be changed to 2503,
however. For why this is read the file!

I changed the word 'historic' to 'historical'. This is because
historical refers to something that happened in the past and the rules
and guidelines are something that existed in the past. On the other hand
historic refers to famous or important in history. Of course that might
be the case here but it felt like in these cases historical was more
correct. This, of course, depends entirely on how you view the rules and
guidelines that are from the previous contest.

This time I modified files in inc: guidelines.closed.hdr and
rules.closed.hdr. This is because the next/guidelines.md and
next/rules.md refer to it as being included (even though it appears that
might not be the case). These files are simply what the top of the
guidelines.md and rules.md files are. I do not know if this is important
but since the markdown files do refer to them I have updated this time
(even though it seems like one has to keep two copies of the same text).