Format fix bugs.md
It seems pandoc actually handles it okay but the markdown was fixed
anyway (in other words it does not change the html file). A line that
was part of a sentence (the end of) started with '256.' This was changed
to '256\.' (another option would be to move the '256.' to the end of the
previous line). With pandoc the html file does not change but this very
issue has caused me problems before so I fixed it.