Fixes in thanks file
The inc/md2html.cfg file was in this case updated for the keywords (not
that many would think of it, maybe, but since I was making another fix I
added the keyword 'ta') and more importantly HEADER_2. The file is not
just about fixing old entries but rather help in every other way too.
Thus it now reads 'Thanking people for their help'.

The thanks-for-help.{md,html} has also been fixed in the above way and
the 'thanks for all the fish' is now a link.

	'Ta ta and thanks for all the fish :-)'