Links and format fixes in bin/README.md
This completes bin/ except that one file is likely to be moved to bin/
which will mean that the file has to be updated again but not in a big
way.

Very few top level html files (including inc/ and some other
subdirectories that are not the years) have to be checked now, I
believe. Some did not require any fixes at all. Still it might be a
couple days before this is done. It is hoped that the entry html files
will go quicker and then the final FAQ and bugs checks can be done.