Move inc/md2html.cfg to bin/
This is a very large commit not only because I forgot to run make www
after updating the menus but I also forgot to do that after the update
to some of the award titles (only quick-readme2index.sh was used).

There is a typo fix in bin/README.md.

This commit does mean that md2html.cfg has to be added to bin/README.md
but that can be done later.

The steps run to verify this works okay:

	make www
	git status
	git diff
	git add -u
	make update
	git status
	git restore $(git check)

The last is important as it is not time for make update. 'git check' is
an alias:

	git diff --name-only