Restore location of all but 5 authors (40 added)
40 of authors which previously had an 'unknown location' were found in
the archive. Only five are left that are unknown.

Rebuilt location.html and authors.html.

Although I did not check every link in these pages I have checked many
and since both files are generated by a script it is probably safe
(enough) to say that these files can be marked as verified to look okay
and have valid links. This means that unless I have missed something,
the top level html files have all been checked except for bugs.html and
faq.html but these have to be done after the entries as they are likely
to be updated as the entries are being looked at.