Many manifest fixes and improvements
Many entry_text fields have been changed, some to make it more
consistent with others (though most likely some slipped through) and
others to make it more correct.

The recent commits which changed filenames are in this update but the
rest of the process will be done soon as part of the same pull request.