Fix link in 1984/mullender/index.html
The link to 'stack' was for abstract data type which not only has no
article but should be call stack. This has been corrected.

Also with this commit the index.html files of 1984 look good. This
includes the 1984/index.html file itself. There is a change in 1986 that
might come before 1985 but this is to be determined later.