Fix make alt in 1988/isaak
The problem is the DATA variable in the Makefile had a file that no
longer exists.

The warning in the Makefile has been corrected too: it should compile
but it won't likely work in modern systems (it is the original but fixed
so that it at least compiles). The isaak.c is the fixed version that
works with both macOS and linux.