Fixes in 1992/westley
The whereami.c file should not have been left in as the scripts added
take care of the problem.

The Makefile now links whereami to westley and whereami.alt to
westley.alt. Running make clobber will remove the links.

The index.html had some mistakes from the time where the C file was
useful. There might still be some mistakes but that will have to be
checked another time.

The manifest has been updated locally but the .entry.json file and the
index.html file have not yet been corrected. This will happen soon after
the rest of the changes of 1992 are committed.