fix 1992/westley/README.md, warn about TAB intending in markdown.md
One must **NOT** use ASCII TAB characters to intent text.  Down the TAB-ed
road is nothing but trouble!  There are _quasi-solutions_, however those
have side effects that lead to a different set of problems.

The `markdown.md` was updated to ask that text not be intended with ASCII tabs.

See [Please do NOT use ASCII tabs to indent
text](https://ioccc-src.github.io/temp-test-ioccc/markdown.html#markdown.html#please-do-not-specify-a-language-for-a-code-block).

We discovered this awhile ago and fixed it in markdown files.  Sorry
(tm Canada 🇨🇦), we forgot to mention this this in `markdown.md`.

BTW: It took some **deep magic** to be able to show an HTML comment
after a starting fenced markdown code block!  See the **NO_COMMENT=**
line in `bin/md2html.cfg`.  :-)

TODO: Stamp out indenting TABs in markdown files, especially in code blocks.

TODO: Consider a rule to detect problematic ASCII TAB intendation
      within makrkdown files.