Make INABIAF format consistent in index.html files
That is to say in the bugs file it will look something like:

	STATUS: INABIAF - please DO NOT fix

but as a html file it is already bold as a heading. This means there is
no need in the index.html file to have the explicit **DO NOT** as it's
in a code block and it is not consistent in what is in bugs.html. For
instance 1984/mullender had:

	STATUS: INABIAF - please **DO NOT** fix

but this meant the index.html file had the '**'s surrounding the DO NOT
which is not consistent with bugs.html.