Improve blockquote styling (.. hopefully :-) )
It is very hard to find a shade of green that works well in all cases
against the white background and as already established blue and purple
cannot work (and some of those would be hard to work with especially
purple unless very dark maybe). Many possibilities were proposed but
they all seem to be a problem so instead of changing the text to green
there is a border on all sides now (forming a box or if you wish a
'block') and the text colour is black. The box / border is dark green.

The text style is still italic but it is unclear if this is wise. On the
one hand quotes are often in italics but with italics in quotes it means
that any quote with italics will not have those italics stand out.

Tested this change in two different forms: a blockquote where there are
blank lines in between the other lines (it's a dialogue) and a back and
forth dialogue broken up between text that is not a blockquote and it
looks good both in macOS Safari and Safari with an iPhone. This test
involves an entry from 1984 (anonymous) that should have been using
quotes instead of code (and now does but will be committed next) and it
does mean that a quick look over the earlier years for code versus
quotes has to be made but that should not take much effort or time.