Rename 2020/ferguson1/deobfuscation -> obfuscation
As the file is not actually code and does not give deobfuscated code but
rather shows how I obfuscated it (indirectly and directly) a better
name, if spoilers will not work, is obfuscation. Thus obfuscation.md and
obfuscation.html are the new names.

Manifest updated; regenerated .entry.json, index.html and other html
files that link to the html file.

In a heading/section (in obfuscation.html) I linked to two files that
are named.