Update guidelines - add new tools description
This is more for those who are interested in how they can run each step
of the mkiocccentry tool as well as how to make it easier to edit a
submission tarball.

Although it looks okay (the rendered html) a more thorough review will
have to be done, tomorrow if possible.