Slight improvement to 7af7a8 (bin/chk-entry.sh)
Instead of using just git use "$GIT_TOOL" as the script already
establishes what git to use.

The comment I added also explains the rationale about this a bit, as it
might not seem immediately clear why it's useful over find(1).

And although it was already tested I forced what would have caused a
failure before, to happen, so that I could verify that this works as I
had originally imagined it: put a file in an entry directory that does
not belong there and then run the script again, to make sure it does not
detect any problem.

Of course with this change one still has to make sure to add valid files
to git but that always was how it was anyway.