reduce bold links in a few link heavy areas
While the use of `**font-weight: bold;` helps links standout in
most cases, there a few "_link heavy_" pages were the **MANY BOLD
LINK IN CLOSE PROXIMITY** looks excessive to the point of being
distracting.  In areas where **nearly everything is BOLD, nothing
seems to stand out**.

Fortunately there seems to be only a very few places where this is a problem:

* `faq.html` - the table of contents
* `authors.html` - nearly the entire page
* `location.html` - nearly the entire page
* `years.html` - nearly the entire page

As an **bold** 😎 experiment we added to the CSS:

```css
/*
 * To make a link non-bold, use: <a class="normal" href="url">link text</a>
 */
a.normal {
    font-weight: normal;
}
```

And then modified select tools and a markdown file to "_unbold_"
those "_link heavy_"  areas of the 4 pages above.

We don't claim that this is the best solution and there may be an
even better way to solve this, nevertheless we **do** think the it
improves the look and feel of the "_link heavy_"  areas of the 4
pages above.   For example, in `faq.html` beyond the table of
contents, the bold links look reasonable (although again there are
some quasi-"_link heavy_" areas down in the page look a bit too
bold, but perhaps not enough to warrant changing so we only altered
the table of contents).

In a few cases, while we were editing code that generated links,
we remove a few "**.**"'s where we didn't think they helped the
look and feel.

Change `bin/gen-authors.sh` VERSION from "1.10.3 2024-05-24"
to "1.10.4 2024-07-13".

Change `bin/gen-location.sh` VERSION from "1.4.2 2024-05-19"
to "1.4.3 2024-07-13".

Change `bin/gen-years.sh` VERSION from "1.3.3 2024-07-07"
to "1.3.4 2024-07-13".

Performed `make www` to accomplish the above and to test under macOS.