Improve FAQ about providing deobfuscated code
It is preferable that unobfuscated code is by the author if they have
(and many have done exactly that).

It would also be preferable to include the author or authors of the entry or
entries if at all possible.

This update includes a link to how to find more information about an
author and I think that FAQ was slightly improved as well.