update and improve make help
Added `make entry2csv` and `make csv2entry` to the help list.

Changed the "syntax" if the lines printed by `make help` so
that when one selects the entire line and pastes that line
into the shell, the trailing comment is not seen as a rule,
nor it is executed.

Made other minor formatting changes of `make help`.