Fix bug status (missing file) in 1993/ant
The unobfuscated version was recently provided by the author so this is
no longer an issue.

A code block in the INABIAF was changed to a quote as it's not code but
something the author wrote.

This entry's index.html might not be done as a question also has to be
addressed before the manifest can be fixed.