Manifest updates for 1993/ant and 2020/endoh2
For 1993/ant it clarifies (even though the extension of the file
probably would suggest it) that the man page is in text format.

For 2020/endoh2 there were some mistakes introduced with the renaming of
spoiler to other things.

Rebuilt the .entry.json and the index.html files for both.

Rebuilt 2020/endoh2/obfuscation/index.html.