improve bin/new-year.sh and Makefile sub-dir processing
The `bin/new-year.sh` now uses make rules to perform some
if its actions.  It modifies the `Makefile` to add the
new year to the `${YEARS}` make variable.

Changed `bin/new-year.sh` VERSION from "1.0 2024-07-21"
to "1.1 2024-07-21".

The top level `Makefile` will test of a Makefile for
a given year exits before calling a make rule for the year.
If there is no `YYYY/Makefile` a warning message is issued

Updated `bin/README.md` to reflect these changes.

Performed `make www` to accomplish the above and to test under macOS.