add bin/jfmt-wrapper.sh and bin/jprint-wrapper.sh
Added wrappers for `jfmt(1)` and `jprint(1)` that
will use `JSONPath.sh` temporarily in the mean time.

FYI: Please do **NOT** be concerted by these tools.  This is a workaround
prior to the **Great Fork Merge** that will allow us to, when jfmt(1)` and
`jprint(1)` exist, easily modify these wrapper tools to use those tools
when they are available.  But as we do **NOT** want to delay IOCCC28,
this workaround will do for now.