Typo fixes in {jfmt,jval}-wrapper.sh scripts
The wrong tools were referred to, likely a yank paste error from
jprint-wrapper.sh (as all three referred to jprint).

These three tools (and jnamval) will be developed (by me and perhaps
partly by Landon) later on, after IOCCC28 (because much more discussion
is needed before they can be developed and because we are trying to get
the next IOCCC running as quickly as possible) so as the comments state,
a different tool is used at this time. The comments, however, note that
it is a hack until the specific tools have been developed and even if it
does not matter that much the right tools should be mentioned.