Minor fixes to FAQ wrt other repos
The FAQ now notes that there can be times where the jparse repo
(https://github.com/xexyl/jparse/) is different from the mkiocccentry
repo's subdirectory jparse. This is one of the reasons the tools in the
mkiocccentry and the jparse repo show the jparse API with the -V
(version) option.

I believe that this is all that's needed for the jparse repo in the FAQ
but time will tell on that.