Add bin/find-invalid-json.sh to validate JSON
This script is run from 'make test' too so that if someone changes a
JSON file and it is invalid the GitHub workflow will report an error.

This script is based on bin/find-missing-links.sh. It can be cleaned up
with exit codes (and consistency checks need to be made) but other
issues have higher priorities so I will leave this the way it is for
now.

Requires verge to be installed and the jparse version to be 1.2.0 or
greater.