Typo fix / updates in thanks file
A typo fix was made which until corrected did not make sense.

Some other minor changes were made, to account for some missing things.

I believe that unless something glaringly bad is noticed, this should be
the last update of this file, unless of course something new comes up,
but if that does I guess it'll be after the great fork merge and
IOCCC28.