change the notice about email to be more generic
We make the note about something changing the subject to be more generic
so that it may both better "stand the test of time" (as the expression
goes) and be more generic to be better "future proof" (as the expression
also goes).

Performed `make quick_www` to update the HTML pages.

TODO: Add images and/oor text blocks to steps in `next/register.md`.