checkpoint add images to next/register.html
We added some images to `next/register.md` and thus to
`next/register.html`.

We modified FAQ 0.0 with these 2 new steps:

3. Wait for your submit server username and initial password
4. Login to the submit server and change your password

This is just a checkpoint: these need to be improved but
we are too tired.

NEEDED TODO: Create a new `next/chpw.html` page and move the images
and text from the new FAQ 0.0 step 4 into this new place.
This will keep the step of FAQ 0.0 shorter to read while
keeping the details and images in a separate web page.